change calculate content_fingerprint
